Added the HorizontalTabbedLayout. It is still underway but it damping my productivity, for now the focus will be the weight in.
The Weight Int (WeighIn) is the focus for now becuase there is still the matter of deciding the frequency of these notifications/reminder. This is crucial.
Another thing about the Scheduling (the reason for the HorizontalTabbedLayout) needs to be studied further. The objective is to make it seamless and joyful to edit and carve out the scheduled exercise periods. There will be a need for something immersive which will involve animations and custom views.